refactor: remove search button from filter bar component
