+++ packages/frontend/src/pages/Inbox/inbox.module.css
@@ -4,6 +4,9 @@
     width: 100%;
 }
 
-.filterBarWrapper {
+.filtersArea {
+    display: flex;
+    column-gap: 1rem;
+    flex-wrap: wrap;
     margin-bottom: 1rem;
 }
